Pass stencil as int32_t so depth or stencil of -1 indicates no data

+++ renderdoc/replay/data_types.h
@@ -206,7 +206,7 @@ struct ModificationValue
 {
 	PixelValue col;
 	float depth;
-	uint8_t stencil;
+	int32_t stencil;
 };
 
 struct PixelModification
